Added the funtion name feature to shell script
Now, the test created with this shell script will have
the function name and the function name is also
be copied in the require statement.
The name of the function is to be given as the second
parameter of the shell script.
first parameter in hte shell script is the name of the file without
any extention. It will automatically put the .js extension.